storing/requirements.txt
pydantic
psycopg2-binary